bug fixing

+++ lib/myTypes/myTypes.h
@@ -3,5 +3,5 @@
 namespace resp32flow
 {
   using temp_t = float;
-  using time_t = int;
+  using time_t = unsigned long;
 }
